[FEAT] 상품 파사드 조건 추가 - 재고보다 주문이 많은 경우 주문을 못한다

+++ src/main/kotlin/kr/hhplus/be/server/domain/product/ProductInfo.kt
@@ -5,7 +5,7 @@ class  ProductInfo{
         val productId: Long,
         val name: String,
         val price: Long,
-        val stock: Long,
+        val stock: Int,
     )
     data class Rank(
         val productId: Long,
